+++ src/test/java/fr/insee/genesis/controller/rest/responses/RawResponseControllerTest.java
@@ -48,6 +48,7 @@ import java.time.ZoneOffset;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
 @Slf4j
 class RawResponseControllerTest {
@@ -122,6 +123,16 @@ class RawResponseControllerTest {
             return rawResponseDataPersistanceStub.findByCampaignIdAndDate(campaignId, startDate, endDate, pageable);
         }
 
+        @Override
+        public long countByCollectionInstrumentId(String collectionInstrumentId) {
+            return 0;
+        }
+
+        @Override
+        public Set<String> getDistinctCollectionInstrumentIds() {
+            return Set.of();
+        }
+        
         @Override
         public Page<RawResponseModel> findRawResponseDataByCollectionInstrumentId(String collectionInstrumentId, Pageable pageable) {
             return null;